Confirm deletion of environment resources

+++ cli/commands/environment/delete.go
@@ -4,12 +4,13 @@ import (
 	"fmt"
 
 	"github.com/sensu/sensu-go/cli"
+	"github.com/sensu/sensu-go/cli/commands/helpers"
 	"github.com/spf13/cobra"
 )
 
 // DeleteCommand adds a command that allows user to delete environments
 func DeleteCommand(cli *cli.SensuCli) *cobra.Command {
-	return &cobra.Command{
+	cmd := cobra.Command{
 		Use:          "delete [ENVIRONMENT]",
 		Short:        "delete specified environment",
 		SilenceUsage: true,
@@ -22,6 +23,13 @@ func DeleteCommand(cli *cli.SensuCli) *cobra.Command {
 
 			org := cli.Config.Organization()
 			env := args[0]
+			if skipConfirm, _ := cmd.Flags().GetBool("skip-confirm"); !skipConfirm {
+				if confirmed := helpers.ConfirmDelete(env, cmd.OutOrStdout()); !confirmed {
+					fmt.Fprintln(cmd.OutOrStdout(), "Canceled")
+					return nil
+				}
+			}
+
 			err := cli.Client.DeleteEnvironment(org, env)
 			if err != nil {
 				return err
@@ -31,4 +39,8 @@ func DeleteCommand(cli *cli.SensuCli) *cobra.Command {
 			return nil
 		},
 	}
+
+	cmd.Flags().Bool("skip-confirm", false, "skip interactive confirmation prompt")
+
+	return &cmd
 }
